requirements.txt
azure-ai-inference
python-dotenv>=1.0.1, <2.0.0
black
ruff